# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "To Do List",
        "version": "1.0"
    },
    "paths": {
        "/api": {
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "Вывод дел",
                "operationId": "60177e6fdadeba441fdb13972e66b160",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "user_id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Помыть посуду"
                                            },
                                            "status": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/delete": {
            "delete": {
                "tags": [
                    "Item"
                ],
                "summary": "Удалить дело",
                "operationId": "25c69f762b33260597c43ae1af2c90f0",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/add": {
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "Добавить дело",
                "operationId": "e3dc4e78fe04a3175bf944c41610141c",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "user_id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Помыть посуду"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "0"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/save": {
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "Изменить дело",
                "operationId": "2e420fec160a3a56e804b6f50aacb34c",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "item_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "user_id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Помыть посуду"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "0"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/status": {
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "Изменить статус",
                "operationId": "0c329866443da39d2de2281a7abd4f3e",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "user_id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Помыть посуду"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "0"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    }
}